backend/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
python-dotenv
aiohttp
google-auth
pandas
numpy
scikit-learn
statsmodels
python-multipart
prophet